jarvis/requirements.txt
# Jarvis Voice Assistant Requirements

# Speech recognition and synthesis
speech_recognition>=3.8.1
pyttsx3>=2.90

# Web and API interactions
requests>=2.28.1
pywhatkit>=5.4
wikipedia>=1.4.0
pyjokes>=0.6.0

# Environment and configuration
python-dotenv>=0.21.0

# AI/LLM integration
ollama>=0.1.0

# For Windows volume control (optional)
# Download NirCmd from https://www.nirsoft.net/utils/nircmd.html
# and place nircmd.exe in your PATH or project directory